move file movement commands before, the ios commands were overriding pushFile and pullFile these commands really belong in file-movement anyway also, fix up the incorrect use of tempDir also, fix up the fact that some clients send in weird byte arrays instead of strings, silly java
